+++ test/pallet/common/deprecate/test_vars.clj
@@ -1,0 +1,4 @@
+(ns pallet.common.deprecate.test-vars)
+
+(def five 5)
+(defn six [] 6)